Create custom SDK build and testing CLI tool
Our previous build process used Gulp + Webpack, which was hard to
understand (webpack config can be complex), slow, not possible to
incrementally build and watch for files over network file systems (for
home development on PC), and prone to a race condition when copying the
resulting SDK bundle to the Rails public/sdks directory where the latest
changes would not be copied over (it became frustrating).

This new build tool takes out Gulp and Webpack (which had quite a few
dependencies and bloated our package.json), and takes advantage of
TypeScript's compiler and a faster module bundler (Brunch) to produce a
leaner build pipeline.

Gulp was never really necessary; the same things could be achieved with
a custom node script, or possibly a shell script, since we were only
copying the SDK bundle and its sourcemaps to a couple Rails project
directories.

Webpack was a great module bundler, and in general is great for complex
frontend projects with stylesheets, image assets, and JavaScript.
However, we are soon removing our stylesheets and serving them
separately; our project now only consists of TypeScript files. For this,
Microsoft's TypeScript compiler has to be used, and we can even skip
using Babel.

How the build process worked before:

- Webpack does its magic, taking all JS files, stylesheets, and images
  (our subscription bell SVG) and turning it into one bundle.
- Gulp would watch for changes to the bundle, and copy it and its
  sourcemaps to a couple Rails project directories.

How our build process works now:

1. TypeScript's compiler turns each TypeScript file into a JavaScript
   file, side by side with each original file
2. Brunch does its magic (all in brunch config, which is easier than
   webpack's):
    2.1 Brunch bundles all the JS files into one module.
    2.2 Brunch bundles all SCSS files into one CSS file.
    2.3 Brunch plugin replaces $_VAR preprocessor variables for things
        like $_VERSION and $_DEV
    2.4 Brunch plugin minifies the CSS using CleanCSS.
    2.5 Brunch plugin minifies the JS using UglifyJS.
3. A source map tool called Sorcery corrects the multi-level source
   mapping (TypeScript source maps --> JS bundle source map --> UglifyJS
   source map).
4. Copy SDK distributable files to Rails project directories public/ and
   public/sdks.

While more lines of code are written for the config, the process is
easier to manage, faster, and supports incremental build over network
file systems.

As a bonus, this CLI tool, being a CLI tool, should provide clear usage
instructions and feedback.

Also, a simple "config.json" sits at the root directory for
configuration of all future build, testing, and other properties.

It can be decked out with future options for different unit and
integration testing.